scons now builds JANA modules correctly

+++ src/plugins/JTest/JTestDataObjects.h
@@ -49,9 +49,9 @@ struct JTestTrackData : public JObject {
 
     void Summarize(JObjectSummary& summary) const override {
         size_t nitems = std::min(buffer.size(), (size_t) 5);
-        for (int i=0; i<nitems; ++i) {
+        for (size_t i=0; i<nitems; ++i) {
             char varname[20];
-            snprintf(varname, 20, "x_%d", i);
+            snprintf(varname, 20, "x_%ld", i);
             summary.add(buffer[i], varname, "%d");
         }
     }
